Отдельный класс для DTO-операций
